+++ test/integration/connectors/elasticsearch/test_elasticsearch.py
@@ -15,7 +15,7 @@ from _pytest.fixtures import TopRequest
 from elasticsearch import Elasticsearch as ElasticsearchClient
 from elasticsearch.helpers import bulk
 
-from test.integration.connectors.utils.constants import DESTINATION_TAG, SOURCE_TAG
+from test.integration.connectors.utils.constants import DESTINATION_TAG, SOURCE_TAG, NOSQL_TAG
 from test.integration.connectors.utils.docker import HealthCheck, container_context
 from test.integration.connectors.utils.validation.source import (
     SourceValidationConfigs,
@@ -177,7 +177,7 @@ def destination_index(elasticsearch_elements_mapping: dict) -> str:
 
 
 @pytest.mark.asyncio
-@pytest.mark.tags(CONNECTOR_TYPE, SOURCE_TAG)
+@pytest.mark.tags(CONNECTOR_TYPE, SOURCE_TAG, NOSQL_TAG)
 async def test_elasticsearch_source(source_index: str, movies_dataframe: pd.DataFrame):
     indexer_config = ElasticsearchIndexerConfig(index_name=source_index)
     with tempfile.TemporaryDirectory() as tempdir:
@@ -207,7 +207,7 @@ async def test_elasticsearch_source(source_index: str, movies_dataframe: pd.Data
         )
 
 
-@pytest.mark.tags(CONNECTOR_TYPE, SOURCE_TAG)
+@pytest.mark.tags(CONNECTOR_TYPE, SOURCE_TAG, NOSQL_TAG)
 def test_elasticsearch_source_precheck_fail_no_cluster():
     indexer_config = ElasticsearchIndexerConfig(index_name="index")
 
@@ -221,7 +221,7 @@ def test_elasticsearch_source_precheck_fail_no_cluster():
         indexer.precheck()
 
 
-@pytest.mark.tags(CONNECTOR_TYPE, SOURCE_TAG)
+@pytest.mark.tags(CONNECTOR_TYPE, SOURCE_TAG, NOSQL_TAG)
 def test_elasticsearch_source_precheck_fail_no_index(source_index: str):
     indexer_config = ElasticsearchIndexerConfig(index_name="index")
 
@@ -236,7 +236,7 @@ def test_elasticsearch_source_precheck_fail_no_index(source_index: str):
 
 
 @pytest.mark.asyncio
-@pytest.mark.tags(CONNECTOR_TYPE, DESTINATION_TAG)
+@pytest.mark.tags(CONNECTOR_TYPE, DESTINATION_TAG, NOSQL_TAG)
 async def test_elasticsearch_destination(
     upload_file: Path,
     destination_index: str,
@@ -282,7 +282,7 @@ async def test_elasticsearch_destination(
         validate_count(client=client, expected_count=expected_count, index_name=destination_index)
 
 
-@pytest.mark.tags(CONNECTOR_TYPE, DESTINATION_TAG)
+@pytest.mark.tags(CONNECTOR_TYPE, DESTINATION_TAG, NOSQL_TAG)
 def test_elasticsearch_destination_precheck_fail():
     connection_config = ElasticsearchConnectionConfig(
         access_config=ElasticsearchAccessConfig(password=ES_PASSWORD),
@@ -297,7 +297,7 @@ def test_elasticsearch_destination_precheck_fail():
         uploader.precheck()
 
 
-@pytest.mark.tags(CONNECTOR_TYPE, DESTINATION_TAG)
+@pytest.mark.tags(CONNECTOR_TYPE, DESTINATION_TAG, NOSQL_TAG)
 def test_elasticsearch_destination_precheck_fail_no_index(destination_index: str):
     connection_config = ElasticsearchConnectionConfig(
         access_config=ElasticsearchAccessConfig(password=ES_PASSWORD),
@@ -312,6 +312,7 @@ def test_elasticsearch_destination_precheck_fail_no_index(destination_index: str
         uploader.precheck()
 
 
+@pytest.mark.tags(CONNECTOR_TYPE, DESTINATION_TAG, NOSQL_TAG)
 @pytest.mark.parametrize("upload_file_str", ["upload_file_ndjson", "upload_file"])
 def test_elasticsearch_stager(
     request: TopRequest,