+++ unstructured_ingest/v2/processes/connectors/databricks/__init__.py
@@ -25,6 +25,8 @@ from .volumes_native import (
     databricks_native_volumes_destination_entry,
     databricks_native_volumes_source_entry,
 )
+from .volumes_table import CONNECTOR_TYPE as VOLUMES_TABLE_CONNECTOR_TYPE
+from .volumes_table import databricks_volumes_delta_tables_destination_entry
 
 add_source_entry(source_type=VOLUMES_AWS_CONNECTOR_TYPE, entry=databricks_aws_volumes_source_entry)
 add_destination_entry(
@@ -49,4 +51,8 @@ add_source_entry(
 )
 add_destination_entry(
     destination_type=VOLUMES_AZURE_CONNECTOR_TYPE, entry=databricks_azure_volumes_destination_entry
+)
+add_destination_entry(
+    destination_type=VOLUMES_TABLE_CONNECTOR_TYPE,
+    entry=databricks_volumes_delta_tables_destination_entry,
 )